// standing-out/StandingOutStore/AngularSource/app/models/admin-tutor-details.ts
export class AdminTutorDetails {
	tutorId: string;
	stripePlanId: string;
	stripeCustomerId: string;
	stripeSubscriptionId: string;
	paymentStatus: string;
	calendarId: number;
	header: string;
	subHeader: string;
	biography: string;
	imageName: string;
	imageDirectory: string;
	hasDbsCheck: boolean;
	dbsCertificateNumber: string;
	dbsApprovalStatus: string;
	profileTeachingExperiance: string;
	profileHowITeach: string;
	profileImageFileLocation: string;
	profileImageFileName: string;


    userFullName: string;
    userFirstName: string;
    userTitle: string;
	storeProfileImageDownload: string;

	stripePlanName: string;
	profileApprovalStatus: string;
	tutorCertificates: any;
}